Updated & Added features to template
Set unregister on reboot to yes
Disabled DHCP Option 42 and 2, this causes problems with timezone, specifically with regards to daylight savings. Since auto is defaulted as an else, leaving this disabled should work fine.
Added DND and Forward features in a way that will play nice with FusionPBX, as well as work well with the Grandstream Phone. The value of having them as IF statements, is the Features button becomes busy with all accounts even if only 1 account is enabled.
May consider adding these to other GRP models.